Made StreamConverters responsible for Steps and Accumulators also.
Split out Accumulators from StepConverters so StepConverters could just be hierarchy of traits for implicit conversion.
